+++ src/dataLayer/dataHelper.ts
@@ -15,3 +15,15 @@ export const postData = (path: string, data: string) => {
     .then((res) => res.json())
     .then((jsonData) => jsonData);
 };
+
+export const deleteData = (path: string, data: any) => {
+  return fetch(`http://localhost:3001/${path}`, {
+    method: 'DELETE',
+    headers: {
+      'Content-Type': 'application/json',
+    },
+    body: JSON.stringify({ data }),
+  })
+    .then((res) => res.json())
+    .then((jsonData) => jsonData);
+};